+++ include/gradido_blockchain_core/data/wire/confirmed_transaction.h
@@ -131,7 +131,7 @@ grd_result grdw_confirmed_transaction_decode(
  */
 grd_result grdw_confirmed_transaction_encode(
     grdu_memory_block *binary_dst,
-    size_t *final_size,
+    int *final_size,
     const grdw_confirmed_transaction *tx,
     grd_memory *allocator
 );
