+++ include/gradido_blockchain_core/data/wire/gradido_transaction.h
@@ -115,7 +115,7 @@ grd_result grdw_gradido_transaction_decode(
  */
 grd_result grdw_gradido_transaction_encode(
     grdu_memory_block *binary_dst,
-    size_t *final_size,
+    int *final_size,
     const grdw_gradido_transaction *tx,
     grd_memory *allocator
 );
